requirements.txt
requests==2.32.3
python-dateutil==2.9.0
inotify==0.2.10
pyyaml==6.0.2